refactor game state
