+++ src/reducers/actor/selectors.ts
@@ -1,6 +1,9 @@
 import IActor from "../../entities/IActor";
 import { IApplicationState } from "../../store/state";
 
+const selectAllActors = (state: IApplicationState): IActor[] =>
+  state.actorState.actorList;
+
 export default {
-  selectAllActors: (state: IApplicationState): IActor[] => state.actor.actorList
+  selectAllActors
 };
