--- src/reducers/game/index.ts
@@ -1,4 +1,0 @@
-export { IGameState } from "./game_state";
-export { default as gameSelectors } from "./game_selectors";
-export { default as gameActions } from "./game_actions";
-export { default } from "./game_reducer";
